Fix vanishing labels for first and last ticks.
Extra space must be reserved by layout for the axes that have wide
labels for first or last tick. Changed the logic how axis sizeHint
is interpreted to make the previously irrelevant height or width
(depending on orientation) of the sizeHint to indicate how far
the widest label extends past the first/last tick, and adjust the
grid size accordingly in layout.

Reviewed-by: Mika Salmela