Fix zooming in crash
The bounding rect has to fit inside int limits when zooming in, meaning
that the height and width both have to be within the int limits
(INT_MAX). QWidget::update() uses a region (based on bounding rect)
that has to be compatible with QRect. The geometry change is only done
if the bounding rect fulfills this requirement.

Task-number: QTRD-1907
Change-Id: I4e874de355390c5fc983ac1e0976bf0d21e2d10c
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@digia.com>