Fix ChartView visibility change drawing
New geometry for the ChartPresenter is calculated only when the chart is
visible. The geometry is set for the layout always.

If the visibility of a series changes then the layout of the legend is
only invalidated in case the chart is visible. This removes unnecessary
delay of the legend drawing when the chart is set visible.

Task-number: QTRD-2064
Change-Id: I78e6da3b859af9becf4f5059e9c02981c79f6d64
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@digia.com>