Fix mouse event handling with RubberBand
The mousePressEvent handles the event differently for rubber band only
if the mouse press occurs on the plot area. This way pressing the mouse
outside plot area will handle the event according to the pressed item
(like legend marker). For the same reason the mouseRelease event handles
the event differently for rubber band only if the rubber band is
displayed.

Task-number: QTRD-2260
Change-Id: Ibedcd2d13a7c2229352074a4df326cd73e9cf80f
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@digia.com>