Fix logarithmic axis point calculation
For logarithmic axis the geometry point is set even if the point value
is zero or negative. For zero and negative values the x-point is set to
zero and the y-point is set to chart height. This way the bar chart
items can ignore the return value from calculateGeometryPoint() method
and not draw the bar for a value that does not have logarithm defined.

Task-number: QTRD-2065
Change-Id: I7ba4efed6a0678eb487377e3e88bdee9c849a9a3
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@digia.com>