Fix crash with ScatterChartItem mouse events
The ScatterChartItem should emit the clicked and hovered signal after
the QGraphicsRectItem has handled the corresponding events. This is to
avoid situations where the item has been changed due to the signal
as it can lead to problems when handling the mouse event (like crashes
due to deleted items).

Task-number: QTRD-2296
Change-Id: Ibdffb158f8a73df6af7fe3868b2707053e8bf975
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@digia.com>