Fix qmlcustomizations example
If the interval coefficient was under 0.25, it resulted as zero
interval for the timer, stopping the animation.

The blinking after animation stopped was also broken, as using var
as a temporary color store doesn't work due to var assignments being
by reference. Using a separate color property instead of var works
around this issue.

Task-number: QTRD-2170
Change-Id: I7224ff259a076da3c6e5e6e66bc0c3031aa1dda6
Reviewed-by: Mika Salmela <mika.salmela@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>