Fix autotests
Hover signal tests:
Hover signals are randomly failing on various platforms when
run on virtual machines, so we skip them for now.
When tested manually with an actual mouse, hover signals work
just fine, but synthetic mouse move generation doesn't seem
to be reliable.

QAreaSeries:
Hide axes so they don't confuse pixel color checks.

QPieSeries:
Removed unnecessary window activation check that had nothing
to do with the test case and was causing it to fail on some
systems.

Changed the button size in mouse event skip test to get rid
of a warning of minimum size in Windows. Also added processEvents
calls to the skip test to minimize impact on the test case.

Change-Id: Ia5914c547cbe58facb3a96c448cbbf3998d4a757
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>