When effective marker is MarkerFromSeries then it can call a pure virtual
When MarkerFromSeries was used it would end up crashing due to a pure
virtual call because the QLegendMarkerPrivate subclass had not finished
being created and therefore it would not be able to call series() to
determine the marker shape.

Therefore the call to updateMarkerShapeAndSize() is removed from the
constructor as it will be called via updated() anyway, and to account
for one possible outcome, added to setGeometry() to ensure the items
are valid in time.

Task-number: QTBUG-65616
Change-Id: Iaa50a83373f570d1d41c48a6184c99d4de2b739b
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>