Implement label editing feature in the QDateTimeAxis
Implement the labelsEditable property handling in the QDateTimeAxis.
If this property set to true the user will be able to manipulate
the axis labels by double clicking on it and entering a new date
in the same format as it is displayed. After finishing the edit
the axis range will be adjusted according to the entered date.

Change-Id: Ide45982875924c8bb4e937b47511dfa3c9237750
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>