Fix access to bar sets instantiated inside the module
The QBarModelMapper allocates QBarSet instances during initialization.
While this works for pure C++ applications, on the QML side these need
to be of type DeclarativeBarSet to be accessible, but also expose
additional properties.
The C++ module is not aware of QML, also the QBarModelMapper cannot
expose a dynamic allocator without breaking binary compatibility.

Hence, the approach is to create a custom allocator function and export
it from the C++ module. In case a declarative chart is created, the
allocator is replaced with the QML one.

Fixes: QTBUG-71013
Change-Id: I1864d637df37f8cd38d28fdba75da0c5608a612c
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>