Place minor tickmarks on horizontal axis correctly
It looks like top and bottom was accidentally swapped when the dynamic
tickmarks feature was added. This resulted in the minor tickmarks being
drawn on the opposite edge of the canvas from the axis and the major
tickmarks.

Also modified the manual test to make drawing the grid optional, so the
position of the tick marks can be seen more easily.

Fixes: QTBUG-79349
Change-Id: Ia25c656d3d66ee0e6dd908504ec5c81cd2555bde
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>