Prevent zoom-out beyond infinite linear max double values
When zooming out domains that include log axes, prevent zoom-out
when the new zoom level would result in a max value that is above
the upper range of double (represented as inf).

Additionally, prevent zoom-out when the log value exceeds the
number of pixels in the log axis dimension. Major gridlines are
created at each integral log value, and it does not make sense to
zoom out beyond such point.

Task-number: QTBUG-75500
Change-Id: Icab72f600a84297d43da675c5757f6e63ae25119
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>