Update the commercial license header to Qt Charts files
The commercial license header header.COMM was updated to
the tqtc-qtcharts files.

The commercial license header may contain some
additional lines so that its line count equals
with the earlier license header. Reason for this
is that some autotests use hard coded line numbers
and a change in the line count causes failures in
tests.

Task-number: QTQAINFRA-4144
Change-Id: I87f9848863d619d3fa0f3b8d94f21c14a87fe869
Reviewed-by: Antti Kokko <antti.kokko@qt.io>
Reviewed-by: Tuukka Turunen <tuukka.turunen@qt.io>