Chart item mouse click handling if chart rubber band enabled
Pass left mouse clicks on to underlying Qt Chart items if those items
accept left mouse clicks, in all cases.

Previous to this change if QChartView's setRubberBand feature was
enabled, left clicks were always handled by the QChartView to start
showing the rubber band selection.

With this change, if the mouse is over a QGraphicsItem in the scene that
accepts Qt::LeftMouse button, then the click is passed on to the item.

To make this work, some QGraphicsItems that are created in the chart
needed to properly mark their acceptance of mouse buttons using
QGraphicsItem::setAcceptedMouseButtons, as the default is to accept all
mouse buttons.

Task-number: QTBUG-94035
Change-Id: I4f2ec3823b10acbb5ef5195454ab5461d04b75da
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>