QXYSeries: Support for selection in a plot
In order to support selection of points in a chart there is a support
for this in scatterchartitem and linechartitem.
The interface for that is available on QXYSeries level.
Selected points can have changed color and they are a bit bigger than
other points.
Selected points are visible regardless of other points visibility.

Task-number: QTBUG-89445
Change-Id: I6d11afa84288f4f949e53b005410659b0ecf3c55
Reviewed-by: Keith Kyzivat <keith.kyzivat@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>