Add move and resize to QLegend
This feature adds the ability to move and resize the chart legend when
it is not attached to the chart. It does not put a separate frame around
the legend, but instead takes control of mouse messages when the mouse
cursor is between the content area and the border of the legend.

The user can drag the legend around when they click and drag the top
edge of a detached legend. A visual rectangle is painted when the user
positions the cursor at the top of a detached legend. This area
identifies the region they can use to drag the legend around.

If the mouse pointer is moved beyond the edge of the chart, it will
attach to the edge that the mouse moved off of. The legend remains bound
to the area of the chart, even if dragging would have moved an edge of
the legend beyond the edge of the chart.

An attached legend can be detached by double clicking it.

Task-number: QTBUG-93477
Change-Id: I5adff45802ba326004f2ffebe27357c8f8b138d0
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>