Fix X axis labels to not draw over each other
When the label numbers on the Y axis become too wide then the plot area
width is shrunk to accommodate for that. But the X axis label numbers
were calculated for the wider X axis, and printed on the shrunk X axis,
so they overlapped each other.

This is addressed in the Qt::AlignBottom part of the fix. Presumably the
same issue happens with different alignments so respective change has
been applied to those too.

Fixes: QTBUG-79218
Change-Id: I2ad8c62f8c8d6308ab13b9da42fa6bd5e9192ee3
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
(cherry picked from commit 5b6e1042bd0aa4514df71209e7784b23490dbd1e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>