Best fit line support added to QXYSeries
Best fit line reflects trend of points data set. This feature allows
user to show and hide such a line on the chart. Look of the line can
be adjusted by setting the pen for best fit line.
Method to get factors (slope and intercept) of the best fit line added.

Task-number: QTBUG-89453
Change-Id: I1c9660e04462352f6b4876e56eebb16b0de83a76
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>