Avoid infinite allocating loop when zooming in too much
Use standard library ceil() and floor() instead of qCeil() and int(),
because the latter return int, and in-between converting to/from
floating-point math, the integer arithmetic causes wrap-arounds.

To be more specific, the following part

  qCeil((minValue - value) / interval)

returns a negative number because of int overflow, when the values are:
minValue = 29.999999972060323
value    = 0.0
interval = 9.3132257461547852e-09

As a result of this fix, the infinite allocating loop is avoided, but
the program terminates with the following Q_ASSERT() when zooming in
towards the limits of floating point precision.

ASSERT: "labels.size() == labelList.size()" in file
    ../qt5/qtcharts/src/charts/axis/horizontalaxis.cpp, line 91

Task-number: QTBUG-93672
Change-Id: I5db17a55c3b10cfcb4e093f0ab78edb49f040a8c
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>