Add optional labels truncation to QAbstractAxis
This is a feature that allows users to switch off labels truncation
that is by default performed if there is too little space for the label.
By default truncateLabels is true, so it doesn't change previous
behavior. If user doesn't want to have truncated labels, he can switch
them off explicitly.

There is also a labelsTruncated method added that returns true if there
is at least one truncated label and false otherwise. The reason for this
is to be able to programmatically check if labels are truncated.

Task-number: QTBUG-89444
Change-Id: I8ab18a8d0335ddf31baf443a2046fee89ce58b38
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>