Display selected points regardless of QXYSeries::lightMarker()
Selected points needs to be displayed regardless of the visibility
of other points. It was correctly implemented when not using light
markers concept, but it didn't work correctly with
QXYSeries::lightMarker() set to null image and valid
QXYSeries::selectedLightMarker().

The fix was to change some of the conditions in linechartitem.cpp,
splinechartitem.cpp and scatterchartitem.cpp. There was also a change
added to scatterchartitem to use QXYSeries::pointsVisible() setting.

Task-number: QTBUG-89445
Pick-to: 6.2
Change-Id: I6010bbf6b6f670f6b0694a7612d22e353d320154
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>