Make sure to install designer files with non-prefixed builds
It is not enough to just call qt_copy_or_install with a DESTINATION
value prefixed with only ${INSTALL_QMLDIR}/ because for non-prefixed
builds this will install relative to the directory this cmake file is
in. Instead we have to combine ${QT_INSTALL_DIR} with the target path to
make sure the final path is always correct. This pattern is used all
over, so we muse do it as well.

Change-Id: I327765dd61224ff070b07a48abc424fd7dadbb37
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit 76c3558507da722f7f523d36dc1cf61fdb376084)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>