Simplify management of two static regular expression objects
ChartAxisElement's methods used two regular expressions to match label
formats. Each was heap-allocated lazily when first needed and there
was a static object whose destructor took care of deleting them.

Repackage that as two little functions, each of which has a local
static instance of QRegularExpression (so it'll be instantiated on the
first call, equivalent to the former lazy construction, and the
compiler / library / runtime can even be expected to take care of any
thread synchronization needed to ensure this is only done by the first
to need it) that it returns.

Change-Id: I61464cddf830e2e077df001bca95d0f962f26571
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Samuel Gaist <samuel.gaist@idiap.ch>