Update commercial license headers
Updated header.COMM to the files in tqtc-qtcharts..
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4929
Change-Id: I52db4ff0f86002ff7a8421cff95337f9868fd9ba
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Antti Kokko <antti.kokko@qt.io>