Fix QtCharts Audio Example not scaling correctly on mobile devices
The minimumSize was set, and this stopped it from scaling correctly on
devices with a smaller screen size in landscape mode.

Added a Widget::resize() call in main.cpp. This replicates the same behavior
as before on desktop platforms and fixes the scaling issue for the small
screen devices.

Fixes: QTBUG-106404
Change-Id: Ifd0395fd85e77f274ae23d59de19278801e6983b
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
(cherry picked from commit f9da051f42f0f21679fae06b14cc10802b2745b7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>