Do not use global pointsConfigurationDirty flag
The flag is toggled only when the pointsConfiguration() is changed.
Later this flag is used to check if we need to resize the markers for
the points.

If the points are removed, and later re-added without changing the
pointsConfiguration(), this flag was preventing a proper size update
of the markers.

Fix it by simply removing the global flag, as all the other places where
pointsConfiguration() is used do not check it anyway.

Task-number: QTBUG-112919
Pick-to: 6.5 6.2
Change-Id: Ic56eb94c94dda319d23f6d6eadfd4df884618064
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Lukasz Kosinski <lukasz@scythe-studio.com>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>