CMake: Make qtcharts tests standalone projects
Add the boilerplate standalone test prelude to each test, so that they
can be opened with an IDE without the qt-cmake-standalone-test script,
but directly with qt-cmake or cmake.

Boilerplate was added using the following scripts:
https://git.qt.io/alcroito/cmake_refactor

Manual adjustments were made where the code was inserted in the wrong
location.

Task-number: QTBUG-93020
Change-Id: I5cc58c1ce3c6670c0cfde06ca640947d647f7fea
Reviewed-by: Amir Masoud Abdol <amir.abdol@qt.io>