Charts: Fix crash when clicking on scatter points
There is a dependency loop in the chart marker implementation that
causes a crash when clicking on the scatter points. The markers hold
pointer to the parent item but when handling the mouseReleasedEvent,
the markers gets deleted and the parent pointer becomes unusable.
The code then returns to the marker code which uses the parent pointer
and crashes.

This change moves the event handling from the marker to the parent item.

Fixes: QTBUG-115072
Pick-to: 6.6 6.5 6.2
Change-Id: Ibdbbdf6b51d8a86f6bae8a9a8b7293ee3760876e
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>