Mark the module as free of Q_FOREACH, except where it isn't
The density of Q_FOREACH uses is exremely high here, too high for this
author, unfamiliar with this code, to tackle in a short amount of
time. Even if they're not concentrated in just a few TUs, we need to
make progress on a global QT_NO_FOREACH default, so grab the nettle
and stick to our strategy:

Mark the whole module with QT_NO_FOREACH, to prevent new uses from
creeping in, and whitelist the affected TUs by #undef'ing
QT_NO_FOREACH locally, at the top of each file. For TUs that are part
of a larger executable, this requires these files to be compiled
separately, so add them to NO_PCH_SOURCES (which implies
NO_UNITY_BUILD_SOURCES, too). Created QTBUG-115824 to keep track of
this, and other such follow-up changes in this module.

Task-number: QTBUG-115824
Change-Id: I7d3d11f5313b14f866557583c2e4e18bba893e58
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>