Mark chartviewer as free of Q_FOREACH, except where it isn't
The density of Q_FOREACH uses is exremely high here, too high for this
author, unfamiliar with this code, to tackle in a short amount of
time. Even if they're not concentrated in just a few TUs, we need to
make progress on a global QT_NO_FOREACH default, so grab the nettle
and stick to our strategy:

Mark the whole test with QT_NO_FOREACH, to prevent new uses from
creeping in, and whitelist the affected TUs by #undef'ing
QT_NO_FOREACH locally, at the top of each file. For TUs that are part
of a larger executable, this requires these files to be compiled
separately, so add them to NO_PCH_SOURCES (which implies
NO_UNITY_BUILD_SOURCES, too). Created QTBUG-115824 to keep track of
this, and other such follow-up changes in this module.

The use in the charts.h header file to be fixed, because uses in
header files makes any .cpp file including the header (in PCH builds:
all) incompatible with QT_NO_FOREACH.

Task-number: QTBUG-115824
Change-Id: I92481adee921665b2e8b755a5778d0b52a39a6c6
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>