GLXYSeriesDataManager: de-pessimize a for loop
We don't need to call QMap::values() to iterate over the list of
values of the map. Iterating over the map itself is equivalent, unless
the loop modifies the map, which it here clearly doesn't do.

Also drop the Q_FOREACH in favor of a normal ranged for loop, which is
safe for the same reason.

Pick-to: 6.6 6.5
Task-number: QTBUG-115824
Change-Id: Iea6261a9d2b50b03836b82edbf9e74285e226093
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>