GLWidget: disconnect during destruction and cleanup
The GLWidget instance cleans up its data structures when the OpenGL
context is about to be destroyed, but the context might be destroyed as
part of destroying the GLWidget (typically when the QOpenGLWidget
destructor runs).
To prevent that the cleanup() function gets called after the GLWidget
destructor has run (which triggers the assert in QObject), disconnect
explicitly.

The crash reproduces with any of the tests using a QAbstractSeries
implementation when the useOpenGL property is set, so add a data-row
to one of them that provokes the assert without the fix.

Fixes: QTBUG-119900
Pick-to: 6.5
Change-Id: Ib4badd0c5df68ed60df6f492cada7954f0cb220a
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
(cherry picked from commit 5c4ddaae03e720cceb853a4d5574fa4755e66aad)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit be3f2005f1d05edf7ecf9f3e5ece6aa4dd1c13a9)