Use formal parameters for signal handlers in the qmloscilloscope example
The usage of a generic parameter name in the signal handler causes the
name collision and this sometimes causes the required functionality not
to be updated in the qml oscilloscope example.

This patch makes sure that we pass formal parameters to the signal
handlers in the qml for this example. Also, removed unused signal
animationsEnabled(bool)

Change-Id: Ie8d058dd7693393b32e96860c3b6340aaaa1c1e6
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 50f5d7be305183ffac728ec7186ea356f68efbdb)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>