Fix duplicate exports in plugins.qmltypes
Between 1.0 and 1.1 version of charts, two qml types changed names,
but were still backed up by the same C++ type. To avoid duplicate
entries for same C++ type in plugins.qmltypes, we just provide two
exports for the same C++ type. This will unify the version numbers
for the combined types as well, but this shouldn't be a problem.

Fixes: QTBUG-115358
Pick-to: 6.9 6.8
Change-Id: I5f8d91fd4902f9f29b589d9f7abb08071105829b
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>