PieSliceAnimation: also interpolate alpha channel
The animation interpolation did not consider the alpha channel when
interpolating QPen and QBrush. As a result, the colors were always
considered opaque, even if the users specified a custom alpha channel.

Fix it by also interpolating the alpha channel.

Apparently, there is no possibility to unit-test the change, since
there's no API to access the PieSliceItem and its properties.

To test the fix, update the manual qmlchartproperties test with the
possibility to change the opacity of the slice for the PieChart.

Amends 7d415c8db78354e069c41bbea0bf9b8a9f20c66a.

Fixes: QTBUG-135240
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: I559067f11ad7fba1853e1af1fe49de228a6eb6f6
Reviewed-by: Sami Varanka <sami.varanka@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>